added types refactoring setting up http
