+++ plugins/duktape/helpers/janus-manager.js
@@ -1,4 +1,9 @@
-/** @param {State} state */
+// @ts-check
+
+/** 
+ * @param {Global} global
+ * @param {State} state 
+ * */
 module.exports = function (global, state) {
 
   var util = require("./util")(global);
@@ -11,7 +16,7 @@ module.exports = function (global, state) {
   /**
    * @param {number} id
    * @param {string} tr
-   * @param {Body} body
+   * @param {IBody} body
    * */
   function handleManagerMessage(id, tr, body) {
     switch (body.request) {
@@ -26,7 +31,7 @@ module.exports = function (global, state) {
   /**
    * @param {number} id
    * @param {string} tr
-   * @param {Body} body
+   * @param {IBody} body
    * */
   function handleJoinManager(id, tr, body) {
     return 1;
@@ -35,7 +40,7 @@ module.exports = function (global, state) {
   /**
    * @param {number} id
    * @param {string} tr
-   * @param {Body} body
+   * @param {IBody} body
    * */
   function handleSyncManager(id, tr, body) {
     return 1;
@@ -54,44 +59,3 @@ module.exports = function (global, state) {
     syncRoomToManager
   }
 }
-
-/**
- * @typedef {{
-  audioCodec: string,
-  display: string,
-  id: number,
-  isConnected: boolean,
-  janusServer: string,
-  private_Id: number,
-  publishers: Array<any>,
-  room: number,
-  subscribers: Array<any>,
-  videoCodec: string,
-  }} PublisherItem
- * */
-
-/** @typedef {{room_id: number, server: string, publisher_list: Array<PublisherItem>}} RoomInfo */
-
-/**
- * @typedef {{
-    sessions: {};
-    tasks: any[];
-    publishers: any[];
-    rooms: {};
-    managerSessions: {};
-    getRoom: (roomId: any) => any;
-    getSession: (sessionID: any) => any;
-    setSession: (session: any) => void;
-    setRoom: (room: any) => void;
-}} State
- */
-
-/**
- * @typedef {{
-       meetingID: number;
-       sessionID: number;
-       handleID: number;
-       request: string;
-       ptype: string;
-     }} Body
- */
