+++ plugins/duktape/helpers/state.js
@@ -1,32 +1,42 @@
+// @ts-check
 
+/**@param {Global} global */
 module.exports = function (global) {
+
+  var util = require("./util")(global);
+
   var state = {
+    janusServer: "webconf.yourcompany.net",
     sessions: {},
     tasks: [],
     publishers: [],
     rooms: {},
     managerSessions: {},
-    getRoom: function(roomId) {
+    getRoom: function (roomId) {
       var room = null;
       if (state.rooms[roomId]) {
         room = state.rooms[roomId];
       } else {
         // new room template
         var newRoomTemplate = { roomId: 0, roomName: "", managerSessionID: 0, publishers: [], sessions: [] };
+        
+        //
+        /* util.startRoomHttpRequest({ server: "localhost", room_id: roomId, publisher_state: [] }); */
+        //
 
         room = newRoomTemplate;
         room.roomId = roomId
         state.rooms[roomId] = room;
       }
-      return room
+      return room;
     },
-    getSession: function(sessionID) {
+    getSession: function (sessionID) {
       var session = null;
       if (state.sessions[sessionID]) {
         session = state.sessions[sessionID];
       } else {
         // Objects Templates
-        var newSessionTemplate = { id: 0, janusServer: janusServer, room: 0, subscribers: [], publishers: [], isConnected: false };
+        var newSessionTemplate = { id: 0, janusServer: state.janusServer, room: 0, subscribers: [], publishers: [], isConnected: false };
         // new session template
         session = newSessionTemplate;
         session.id = sessionID;
@@ -35,11 +45,11 @@ module.exports = function (global) {
       }
       return session
     },
-    setSession: function(session) {
+    setSession: function (session) {
       state.sessions[session.id] = session;
       console.log("session (" + session.id + ") was updated!!!! ", session);
     },
-    setRoom: function(room) {
+    setRoom: function (room) {
       state.rooms[room.roomId] = room;
     },
   }
