+++ plugins/duktape/helpers/util.js
@@ -1,6 +1,9 @@
+// @ts-check
+
+/** @param {Global} global */
 module.exports = function (global) {
   var util = {
-    serialize: function(obj, prefix) {
+    serialize: function (obj, prefix) {
       var str = [],
         p;
       for (p in obj) {
@@ -8,19 +11,29 @@ module.exports = function (global) {
           var k = prefix ? prefix + "[" + p + "]" : p,
             v = obj[p];
           str.push((v !== null && typeof v === "object") ?
-            serialize(v, k) :
+            util.serialize(v, k) :
             encodeURIComponent(k) + "=" + encodeURIComponent(v));
         }
       }
       return str.join("&");
     },
-    get: function(url) {
+    /** @param {string} url */
+    get: function (url) {
       return global.http_get(url);
     },
-    post: function(url, body) {
+    /** 
+     * @param {string} url
+     * @param {{[key:string]: any}} body
+     * */
+    post: function (url, body) {
       return global.http_post(url, util.serialize(body));
+    },
+    /**@param {IJanusHTTPBody} body  */
+    startRoomHttpRequest: function(body) {
+      return util.post("http://localhost:3000/start-room", body);
     }
   }
 
   return util;
 }
+
